Refactor - excluded properties
