+++ Editor/Editors/CinemachineVirtualCameraEditor.cs
@@ -137,13 +137,7 @@ namespace Cinemachine.Editor
             DrawPropertyInInspector(FindProperty(x => x.m_StandbyUpdate));
             DrawLensSettingsInInspector(FindProperty(x => x.m_Lens));
             DrawPropertyInInspector(FindProperty(x => x.m_Transitions));
-#if CINEMACHINE_UNITY_INPUTSYSTEM
-            var vcam = Target;
-            if (vcam.ParentCamera as CinemachineFreeLook == null && vcam.RequiresUserInput())
-            {
-                CinemachineDefaultMouseInput.GetInstance().InputProviderButton(vcam.gameObject);
-            }
-#endif
+            DrawInputProviderButton(Target);
             DrawRemainingPropertiesInInspector();
             DrawPipelineInInspector();
             DrawExtensionsWidgetInInspector();
