+++ Editor/Experimental/CinemachineNewFreeLookEditor.cs
@@ -55,13 +55,7 @@ namespace Cinemachine
             DrawPropertyInInspector(FindProperty(x => x.m_StandbyUpdate));
             DrawLensSettingsInInspector(FindProperty(x => x.m_Lens));
             DrawPropertyInInspector(FindProperty(x => x.m_Transitions));
-#if CINEMACHINE_UNITY_INPUTSYSTEM
-            var vcam = Target;
-            if (vcam.RequiresUserInput())
-            {
-                CinemachineDefaultMouseInput.GetInstance().InputProviderButton(vcam.gameObject);
-            }
-#endif
+            DrawInputProviderButton(Target);
             DrawRemainingPropertiesInInspector();
 
             // Orbits
