+++ Editor/Experimental/CinemachineNewVirtualCameraEditor.cs
@@ -34,13 +34,7 @@ namespace Cinemachine
             DrawPropertyInInspector(FindProperty(x => x.m_StandbyUpdate));
             DrawLensSettingsInInspector(FindProperty(x => x.m_Lens));
             DrawPropertyInInspector(FindProperty(x => x.m_Transitions));
-#if CINEMACHINE_UNITY_INPUTSYSTEM
-            var vcam = Target;
-            if (vcam.RequiresUserInput())
-            {
-                CinemachineDefaultMouseInput.GetInstance().InputProviderButton(vcam.gameObject);
-            }
-#endif
+            DrawInputProviderButton(Target);
             DrawRemainingPropertiesInInspector();
 
             // Pipeline Stages
