--- Editor/Helpers/CinemachineDefaultMouseInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 40f9d8cd96ddc4d66bc940162f14763f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
